+++ cli/tests/pcluster/cli/test_export_cluster_logs.py
@@ -41,6 +41,7 @@ class TestExportClusterLogsCommand:
         [
             ({"filters": ["Name=wrong,Value=test"]}, "filters parameter must be in the form"),
             ({"filters": ["private-dns-name=test"]}, "filters parameter must be in the form"),
+            ({"filters": "private-dns-name=test"}, "filters parameter must be in the form"),
         ],
     )
     def test_invalid_args(self, args, error_message, run_cli, capsys):
@@ -116,6 +117,7 @@ class TestExportClusterLogsCommand:
                 "output_file": args.get("output_file") and os.path.realpath(args.get("output_file")),
                 "start_time": args.get("start_time") and to_utc_datetime(args["start_time"]),
                 "end_time": args.get("end_time") and to_utc_datetime(args["end_time"]),
+                "filters": [args.get("filters")] if args.get("filters") else None,
             }
         )
         export_logs_mock.assert_called_with(**expected_params)